+++ CHANGELOG.md
@@ -5,7 +5,7 @@
     - inlining loop calculation where possible
     - store node states in objects instead of globs
     - refactored entirely NodeImplementation
-- objects `[tabread~]`, `[tabread~]`, `[symbol]`
+- objects `[tabread~]`, `[tabread4~]`, `[symbol]`
 - in CLI : renamed `appTemplate` format to `app`, `compiledJs` to `javascript` and `compiledAsc` to `assemblyscript`
 
 1.0.0-alpha.4